requirements.txt
requests
beautifulsoup4
pyyaml
colorama
validators
cryptography
dnspython
flask
flask-socketio

